Allow reversing URLs with multiple operations.
A pretty glaring shortcoming is that it is impossible to reverse URLs
that have been registered with same path, but different operations.

Using the same path for different operations is a pretty common pattern.
It is even used throughout Ninja's documentation.

    @router.get("/{identifier}")
    def detail(request, identifier: str):
        return "ok"

    @router.put("/{identifier}")
    def update(request, identifier: str, data: Policy):
        return "ok"

The above is perfectly valid, but reversing `update` using
`django.urls.reverse()` will result in a `NoReverseMatch` error.

This commit is an attempt at solving the issue. N.B. I am not familiar
with the internals of Ninja. This is just an experiment.